+++ waspc/cli/src/Wasp/Cli/Command/Start/Db.hs
@@ -8,7 +8,7 @@ import Control.Monad (when)
 import qualified Control.Monad.Except as E
 import Control.Monad.IO.Class (liftIO)
 import Data.Function ((&))
-import Data.Maybe (fromMaybe, isJust)
+import Data.Maybe (isJust)
 import qualified Options.Applicative as Opt
 import StrongPath (Abs, Dir, File', Path', Rel, fromRelFile)
 import System.Environment (lookupEnv)
@@ -24,18 +24,15 @@ import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require (InWaspProject (InWaspProject), require)
 import Wasp.Cli.Util.Parser (parseArguments)
+import Wasp.Db.Postgres (defaultDockerImageForPostgres)
 import qualified Wasp.Message as Msg
 import Wasp.Project.Common (WaspProjectDir, makeAppUniqueId)
 import Wasp.Project.Db (databaseUrlEnvVarName)
 import qualified Wasp.Project.Db.Dev.Postgres as Dev.Postgres
 import Wasp.Project.Env (dotEnvServer)
 import Wasp.Util (whenM)
+import Wasp.Util.Docker (DockerImageName)
 import qualified Wasp.Util.Network.Socket as Socket
-
--- | Command-line arguments for `wasp start db`
-data StartDbArgs = StartDbArgs
-  { customDbImage :: Maybe String
-  }
 
 -- | Starts a "managed" dev database, where "managed" means that
 -- Wasp creates it and connects the Wasp app with it.
@@ -53,24 +50,29 @@ start args = do
   throwIfCustomDbAlreadyInUse appSpec
 
   let (appName, _) = ASV.getApp appSpec
+
   case ASV.getValidDbSystem appSpec of
     AS.App.Db.SQLite -> noteSQLiteDoesntNeedStart
-    AS.App.Db.PostgreSQL -> startPostgreDevDb waspProjectDir appName (customDbImage startDbArgs)
+    AS.App.Db.PostgreSQL -> startPostgresDevDb waspProjectDir appName (dbImage startDbArgs)
   where
     noteSQLiteDoesntNeedStart =
       cliSendMessageC . Msg.Info $
         "Nothing to do! You are all good, you are using SQLite which doesn't need to be started."
 
--- | Parser for `wasp start db` arguments
 startDbArgsParser :: Opt.Parser StartDbArgs
 startDbArgsParser =
   StartDbArgs
-    <$> Opt.optional
-      ( Opt.strOption $
-          Opt.long "image"
-            <> Opt.metavar "IMAGE"
-            <> Opt.help "Docker image to use for the database (default: postgres)"
+    <$> Opt.strOption
+      ( Opt.long "db-image"
+          <> Opt.metavar "IMAGE"
+          <> Opt.help "Docker image to use for the database"
+          <> Opt.showDefault
+          <> Opt.value defaultDockerImageForPostgres
       )
+
+data StartDbArgs = StartDbArgs
+  { dbImage :: DockerImageName
+  }
 
 throwIfCustomDbAlreadyInUse :: AS.AppSpec -> Command ()
 throwIfCustomDbAlreadyInUse spec = do
@@ -106,31 +108,25 @@ throwIfCustomDbAlreadyInUse spec = do
     throwCustomDbAlreadyInUseError msg =
       E.throwError $ CommandError "You are using custom database already" msg
 
-startPostgreDevDb :: Path' Abs (Dir WaspProjectDir) -> String -> Maybe String -> Command ()
-startPostgreDevDb waspProjectDir appName customImage = do
+startPostgresDevDb :: Path' Abs (Dir WaspProjectDir) -> String -> String -> Command ()
+startPostgresDevDb waspProjectDir appName dbDockerImage = do
   throwIfExeIsNotAvailable
     "docker"
     "To run PostgreSQL dev database, Wasp needs `docker` installed and in PATH."
   throwIfDevDbPortIsAlreadyInUse
 
-  let dockerImage = fromMaybe "postgres" customImage
-
   cliSendMessageC . Msg.Info $
-    unlines $
+    unlines
       [ "✨ Starting a PostgreSQL dev database (based on your Wasp config) ✨",
-        ""
+        "",
+        "Additional info:",
+        " ℹ Using Docker image: " <> dbDockerImage,
+        " ℹ Connection URL, in case you might want to connect with external tools:",
+        "     " <> connectionUrl,
+        " ℹ Database data is persisted in a docker volume with the following name"
+          <> " (useful to know if you will want to delete it at some point):",
+        "     " <> dockerVolumeName
       ]
-        ++ ( if isJust customImage
-               then [" ℹ Using custom Docker image: " <> dockerImage, ""]
-               else []
-           )
-        ++ [ "Additional info:",
-             " ℹ Connection URL, in case you might want to connect with external tools:",
-             "     " <> connectionUrl,
-             " ℹ Database data is persisted in a docker volume with the following name"
-               <> " (useful to know if you will want to delete it at some point):",
-             "     " <> dockerVolumeName
-           ]
 
   cliSendMessageC $ Msg.Info "..."
 
@@ -157,7 +153,7 @@ startPostgreDevDb waspProjectDir appName customImage = do
           Dev.Postgres.defaultDevPass
           Dev.Postgres.defaultDevUser
           dbName
-          dockerImage
+          dbDockerImage
   liftIO $ callCommand command
   where
     dockerVolumeName = makeWaspDevDbDockerVolumeName waspProjectDir appName