+++ waspc/src/Wasp/Util/Docker.hs
@@ -1,0 +1,3 @@
+module Wasp.Util.Docker (DockerImageName) where
+
+type DockerImageName = String